feat: improved the breakpoints order on admin panel
