+++ packages/editor/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+### Improvements
+- Reorganized the breakpoint device sizes in the settings panel to be more intuitive and easier to understand.
+
 ### Bug Fixes
 - Fixed an issue where styles weren't being applied in the correct order across different screen sizes, ensuring your responsive designs now work as expected.
